Create Problem: 185. Department Top Three Salaries
Solution to problem 185 - Department Top Three Salaries.

A company's executives are interested in seeing who earns the most money in each of the company's departments. A high earner in a department is an employee who has a salary in the top three unique salaries for that department.

Write a solution to find the employees who are high earners in each of the departments.